+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 * Added user preferences
 * Added group size/time properties
+* Added support for analyzing all compiled Editor assemblies
 * Fixed Diagnostic Rules serialization
 * Fixed *Home* page *NullReferenceException* on Build
 * Improved issue creation code-readability by using *ProjectIssueBuilder*